Moved GatewayGenerator into Gateways Package
Moved the GatewayGenerator class to another package.
